backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.1
python-dotenv==1.0.1
python-multipart==0.0.9
